Change PR Title check
Instead of only checking the PR title we'll now check the title and all
commit message headlines. At least one of them needs to conform to
conventional commits (cc).

The goal is to avoid duplicating messages in the changelog, as the
commit might have already be marked as a feature and now the PR title
also needs to conform to cc and badaboom badabang, you've got yourself a
duplicate.

In addition this allows cc messages to be more precisely placed.